Add GPU landmarks metric; Sort GPU cost functions
