+++ src/deform_lib/CMakeLists.txt
@@ -59,8 +59,17 @@ if (DF_USE_CUDA)
         "filters/gpu/resample.cu"
         "filters/gpu/resample.h"
 
-        "registration/gpu/cost_function.cu"
-        "registration/gpu/cost_function.h"
+        "registration/gpu/cost_functions/binary_function.cu"
+        "registration/gpu/cost_functions/binary_function.h"
+        "registration/gpu/cost_functions/cross_correlation.cu"
+        "registration/gpu/cost_functions/cross_correlation.h"
+        "registration/gpu/cost_functions/landmarks.cu"
+        "registration/gpu/cost_functions/landmarks.h"
+        "registration/gpu/cost_functions/squared_distance.cu"
+        "registration/gpu/cost_functions/squared_distance.h"
+        "registration/gpu/cost_functions/sub_function.h"
+        "registration/gpu/cost_functions/unary_function.h"
+
         "registration/gpu/transform.cu"
         "registration/gpu/transform.h"
         "registration/gpu_registration_engine.cpp"
