+++ src/deform_lib/registration/gpu/cost_functions/cost_function.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "sub_function.h"
+#include "unary_function.h"
+#include "binary_function.h"
+#include "cross_correlation.h"
+#include "landmarks.h"
+#include "squared_distance.h"
+
