+++ src/deform_lib/registration/gpu/cost_functions/landmarks.h
@@ -1,0 +1,29 @@
+#pragma once
+
+#include "sub_function.h"
+
+#include <thrust/device_vector.h>
+
+struct GpuCostFunction_Landmarks : public GpuSubFunction
+{
+    GpuCostFunction_Landmarks(const std::vector<float3>& fixed_landmarks,
+                              const std::vector<float3>& moving_landmarks,
+                              const stk::GpuVolume& fixed);
+
+    ~GpuCostFunction_Landmarks();
+
+    void cost(
+        stk::GpuVolume& df,
+        const float3& delta,
+        float weight,
+        const int3& offset,
+        const int3& dims,
+        stk::GpuVolume& cost_acc,
+        stk::cuda::Stream& stream
+    );
+
+    const thrust::device_vector<float3> _landmarks;
+    thrust::device_vector<float3> _displacements;
+    const stk::GpuVolume _fixed;
+};
+
